+++ src/main/java/com/dku/council/domain/bus/repository/impl/BusArrivalRedisRepository.java
@@ -3,8 +3,8 @@ package com.dku.council.domain.bus.repository.impl;
 import com.dku.council.domain.bus.model.CachedBusArrivals;
 import com.dku.council.domain.bus.repository.BusArrivalRepository;
 import com.dku.council.global.base.AbstractKeyValueCacheRepository;
-import com.dku.council.global.component.model.CacheObject;
 import com.dku.council.global.config.redis.RedisKeys;
+import com.dku.council.global.model.CacheObject;
 import com.dku.council.infra.bus.model.BusArrival;
 import com.fasterxml.jackson.databind.JavaType;
 import com.fasterxml.jackson.databind.ObjectMapper;
