+++ src/main/java/com/dku/council/global/interceptor/VoidSuccessResponseInterceptor.java
@@ -1,6 +1,6 @@
 package com.dku.council.global.interceptor;
 
-import com.dku.council.global.dto.ResponseSuccessDto;
+import com.dku.council.global.model.dto.ResponseSuccessDto;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
