+++ src/main/java/com/dku/council/global/model/dto/ResponseIdDto.java
@@ -1,4 +1,4 @@
-package com.dku.council.global.dto;
+package com.dku.council.global.model.dto;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
