+++ src/test/java/com/dku/council/domain/rental/controller/RentalControllerTest.java
@@ -11,7 +11,7 @@ import com.dku.council.domain.user.model.entity.Major;
 import com.dku.council.domain.user.model.entity.User;
 import com.dku.council.domain.user.repository.MajorRepository;
 import com.dku.council.domain.user.repository.UserRepository;
-import com.dku.council.global.dto.ResponseIdDto;
+import com.dku.council.global.model.dto.ResponseIdDto;
 import com.dku.council.mock.MajorMock;
 import com.dku.council.mock.RentalItemMock;
 import com.dku.council.mock.RentalMock;
